Fixing a build error for 64 bit (also removed an unneeded dependency) for pyFSD
[git-p4: depot-paths = "//eve/branches/development/MAINLINE/eve/src/pyFSD/": change = 1053505]

+++ FsdObject.cpp
@@ -57,7 +57,7 @@ void FsdObject::SetObjectData(const char* data, uint32_t offset, const FsdSchema
 	}
 
 	uint32_t offsetAttributeArrayStart = m_offset + m_objectSchemaAttributes.endOfFixedSizedData + 8;
-	uint32_t sizeOfOffsetAttributeTable = 4 * attributesWithVariableOffset.size();
+	uint32_t sizeOfOffsetAttributeTable = (uint32_t)(4 * attributesWithVariableOffset.size());
 
 	m_offsetToVariableSizedData = offsetAttributeArrayStart + sizeOfOffsetAttributeTable;
